Downloads/my-flask-tensorflow-app/requirements.txt
Flask==2.1.1
tensorflow==2.14.0
numpy==1.23.0
pillow==9.0.0
flask-cors==3.1.1
gunicorn==20.1.0
